requirements.txt
TA-Lib
backtrader
flask
numpy
python-binance
websocket_client